remove welcome page wrapper
