+++ src/router/index.ts
@@ -9,7 +9,7 @@ import Redirect from "../views/errors/403.vue";
 import ProfileWrapper from "../components/profile/Wrapper.vue";
 import Settings from "../views/common/Settings.vue";
 import PageNotFound from "../views/errors/404.vue";
-import WelcomePage from "../views/common/WelcomePageWrapper.vue";
+import WelcomePage from "../views/common/Welcome.vue";
 import AboutPage from "../views/common/About.vue";
 
 const routerHistory = createWebHistory(process.env.BASE_URL);
